Record usage of rewards
Multiple reward usage is recorded and displayed on the history tab.
Still need to find a way to handle toggling the background of selected
items once the CAB is enabled.

Also noticed that adding chores and rewards don't automatically update
the history tab like it should.  Need to figure out why that is.

The order on the history tab isn't right either; probably need to change
the data type for the created at column.